backend/pyproject.toml
[project]
name = "code-refinery-backend"
version = "1.0.0"
description = "AI-powered code refactoring service backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.12",
    "uvicorn[standard]==0.34.3",
    "pydantic==2.11.5",
    "python-dotenv==1.1.0",
    "httpx==0.28.1",
    "openai==1.86.0",
    "anthropic==0.54.0",
    "google-generativeai==0.8.5",
    "python-multipart==0.0.20",
]

[project.optional-dependencies]
test = [
    "pytest==8.4.0",
    "pytest-asyncio==1.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v" 